task 1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -10,8 +10,8 @@
 
 unsigned long int _power(unsigned int base, unsigned int pow)
 {
-	unsigned int num;
-	unsigned int d;
+	unsigned long int num;
+	unsigned long int d;
 
 	num = 1;
 	for (d = 1; d <= pow; d++)
